+++ app/models/nominatim_result.py
@@ -1,9 +1,0 @@
-from typing import NamedTuple
-
-from shapely import Point, Polygon
-
-
-class NominatimResult(NamedTuple):
-    point: Point
-    name: str
-    bounds: Polygon