+++ tests/lib/test_feature_name.py
@@ -1,20 +1,8 @@
 import pytest
 
-from app.lib.feature_name import feature_name, feature_prefix
+from app.lib.feature_name import feature_name, features_prefixes
 from app.lib.translation import translation_context
-
-
-@pytest.mark.parametrize(
-    ('type', 'tags', 'expected'),
-    [
-        ('way', {'boundary': 'administrative', 'admin_level': '2'}, 'Country Boundary'),
-        ('node', {'amenity': 'restaurant'}, 'Restaurant'),
-        ('node', {}, 'Node'),
-    ],
-)
-def test_feature_prefix(type, tags, expected):
-    with translation_context('en'):
-        assert feature_prefix(type, tags) == expected
+from app.models.db.element import Element
 
 
 @pytest.mark.parametrize(
@@ -30,3 +18,26 @@ def test_feature_prefix(type, tags, expected):
 def test_feature_name(tags, expected):
     with translation_context('pl'):
         assert feature_name(tags) == expected
+
+
+@pytest.mark.parametrize(
+    ('type', 'tags', 'expected'),
+    [
+        ('way', {'boundary': 'administrative', 'admin_level': '2'}, 'Country Boundary'),
+        ('node', {'amenity': 'restaurant'}, 'Restaurant'),
+        ('node', {}, 'Node'),
+    ],
+)
+def test_features_prefixes(type, tags, expected):
+    element = Element(
+        changeset_id=1,
+        type=type,
+        id=1,
+        version=1,
+        visible=False,
+        tags=tags,
+        point=None,
+        members=[],
+    )
+    with translation_context('en'):
+        assert features_prefixes((element,)) == (expected,)